+++ src/main/java/com/pro/WOLmgr/controller/DeviceController.java
@@ -1,6 +1,7 @@
 package com.pro.WOLmgr.controller;
 
 import com.pro.WOLmgr.dto.DeviceRequestDTO;
+import com.pro.WOLmgr.dto.DeviceResponseDTO;
 import com.pro.WOLmgr.service.DeviceService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -16,21 +17,27 @@ public class DeviceController {
     private final DeviceService deviceService;
 
     @PostMapping("/device")
-    public ResponseEntity<?> deviceRegister(@RequestBody DeviceRequestDTO deviceRequestDTO){
-        if(!deviceRequestDTO.getDeviceName().matches("[a-zA-Z0-9]+")){
-            return new ResponseEntity<>("Only English and numbers are available.", HttpStatus.BAD_REQUEST);
-        }else if(deviceService.deviceNameCheck(deviceRequestDTO.getDeviceName())){
-            return new ResponseEntity<>("Duplicate names are not allowed.",HttpStatus.BAD_REQUEST);
-        }else if(deviceService.macAddressCheck(deviceRequestDTO.getIpAddress())){
-            return new ResponseEntity<>("Duplicate IP addresses are not allowed.",HttpStatus.BAD_REQUEST);
-        }
+    public ResponseEntity<DeviceResponseDTO> deviceRegister(@RequestBody DeviceRequestDTO deviceRequestDTO){
+//        TODO: 뷰단에서 자바스크립트 코드로 검증하도록 하자.
+//        if(!deviceRequestDTO.getDeviceName().matches("[a-zA-Z0-9]+")){
+//            return new ResponseEntity<>("Only English and numbers are available.",HttpStatus.BAD_REQUEST);
+//        }
         return new ResponseEntity<>(deviceService.register(deviceRequestDTO), HttpStatus.OK);
     }
 
+    @GetMapping("/device/duplication/{deviceName}")
+    public ResponseEntity<Boolean> deviceNameRead(@PathVariable String deviceName){
+        return new ResponseEntity<>(deviceService.deviceNameCheck(deviceName),HttpStatus.OK);
+    }
+
+    @GetMapping("/device/duplication/{macAddress}")
+    public ResponseEntity<Boolean> deviceMacAddressRead(@PathVariable String macAddress){
+        return new ResponseEntity<>(deviceService.macAddressCheck(macAddress),HttpStatus.OK);
+    }
+
     @DeleteMapping("/device/{deviceNumber}")
-    public ResponseEntity<?> deviceDelete(@PathVariable Long deviceNumber){
-        return deviceService.delete(deviceNumber)?
-                new ResponseEntity<>("Deletion failed.",HttpStatus.INTERNAL_SERVER_ERROR):
-                new ResponseEntity<>(deviceNumber,HttpStatus.NO_CONTENT);
+    public ResponseEntity<Long> deviceDelete(@PathVariable Long deviceNumber){
+        deviceService.delete(deviceNumber);
+        return new ResponseEntity<>(deviceNumber,HttpStatus.NO_CONTENT);
     }
 }
